requirements.txt
# Core Requirements
gunicorn==21.2.0
flask==3.0.2
python-dotenv>=1.0.0
openpyxl>=3.1.0

# AI/ML Components
pandasai==3.0.0b15
langchain==0.1.14
langchain-community==0.0.30
langchain-core==0.1.37
langchain-text-splitters==0.0.1
mistralai==0.4.2

# Markdown Processing
markdown==3.6
tabulate

# Common dependencies
typing-extensions==4.12.0
pydantic==2.9.0
pandas==2.2.2 